emp-registration/requirements.txt
mysql.connector == 2.2.9
mysql-connector-python==8.0.31
